Fix disassemble method

+++ include/NES/cpu6502.h
@@ -163,7 +163,7 @@ class Cpu6502 {
     uint16_t getPC();
     uint8_t getAcc();
     bool isComplete();
-    std::map<uint16_t, std::string> disassemble(uint16_t start, uint16_t end);
+    std::map<uint16_t, std::string> disassemble(uint32_t start, uint32_t end);
 };
 
 #endif /* CPU6502_h */
